move cache & DB logic to Services
